+++ driver/src/main/java/oracle/nosql/driver/values/NullValue.java
@@ -44,6 +44,11 @@ public class NullValue extends FieldValue {
     }
 
     @Override
+    public String getString() {
+        return toJson(null);
+    }
+
+    @Override
     public String toJson(JsonOptions options) {
         return "null";
     }